fix(mysql): prevent precision loss in DECIMAL and NUMERIC types (#695)

+++ utils/testutils/test_schema.go
@@ -28,8 +28,8 @@ var GlobalTypeMapping = map[string]string{
 	// Floating Point Types
 	"float":   "float",
 	"real":    "float",
-	"decimal": "float",
-	"numeric": "float",
+	"decimal": "double",
+	"numeric": "double",
 	"float4":  "float",
 	"money":   "float",
